Allow root scope to belong to activity.
Changes that allow the root scope to be at the activity level, instead of
forcing there to be a level above that.

Sample app now has no base activity class, and almost managed to get rid of
its custom application class.

Extracts a MainView out of the MainActivity, hopefully a step toward having
both mobile and tablet layouts.