Proposal for removing getMortarScope()
I always hated that each android view has to implement getMortarScope which in turns redelegates to Mortar based on the context.. so I figured a way to remove the need for "getMortarScope()" on views, and replace it with a need for "getContext()", which views conveniently already have.

The old ViewPresenter has been renamed to ScopedViewPresenter, it's only needed when your view wants to provide a custom scope.