Getting rid of MainView
This is an experiment.

This makes DemoActivity the view bound to Main.Presenter. No need for a custom view, can use a good old FrameLayout.

Also moves DemoActivity injects where it should be (Main module).

The FlowOwner now requires a MortarContext rather than a View.

We could provide either way by creating a ContextFlowOwner and a ViewFlowOwner