Kill MortarContext in favor of getSystemService()
Warning: This is an API-breaking change.

Casting made it difficult to interoperate with other context wrappers,
since Mortar's would always need to be outermost.  We still keep the
name of the system service lookup private.

Activities have a little more work to do, but not much.  It may or may
not prove worthwhile to introduce helpers for that work later.